// apps/web/generated/api/models/portfolioItemResponse.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Freelance Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface PortfolioItemResponse {
  item_id: string;
  freelancer_profile_id: string;
  title: string;
  description: string | null;
  external_url: string | null;
  file_asset_id: string | null;
  display_order: number;
  is_featured: boolean;
}